+++ hotrod-project/hotrod-generator/src/main/java/org/hotrod/config/SelectGenerationTag.java
@@ -73,13 +73,13 @@ public class SelectGenerationTag extends AbstractConfigurationTag {
 
   @XmlAttribute(name = "strategy")
   public void setSStrategy(final String strategy) {
-    log.info("strategy=" + strategy);
+    log.debug("strategy=" + strategy);
     this.sStrategy = strategy;
   }
 
   @XmlAttribute(name = "temp-view-base-name")
   public void setTempViewBaseName(final String tempViewBaseName) {
-    log.info("tempViewBaseName=" + tempViewBaseName);
+    log.debug("tempViewBaseName=" + tempViewBaseName);
     this.tempViewBaseName = tempViewBaseName;
   }
 
@@ -89,7 +89,7 @@ public class SelectGenerationTag extends AbstractConfigurationTag {
 
     // strategy
 
-    log.info("this.sStrategy=" + this.sStrategy);
+    log.debug("this.sStrategy=" + this.sStrategy);
     if (this.sStrategy == null) {
       this.strategy = DEFAULT_STRATEGY;
     } else {
@@ -101,7 +101,7 @@ public class SelectGenerationTag extends AbstractConfigurationTag {
                 + Stream.of(SelectStrategy.values()).map(s -> s.getCaption()).collect(Collectors.joining(", ")));
       }
     }
-    log.info("this.strategy=" + this.strategy);
+    log.debug("this.strategy=" + this.strategy);
 
     // temp-view-base-name