+++ hotrod-project/hotrod-generator/src/main/java/org/hotrod/exceptions/CouldNotResolveNameException.java
@@ -1,0 +1,11 @@
+package org.hotrod.exceptions;
+
+public class CouldNotResolveNameException extends Exception {
+
+  private static final long serialVersionUID = 1L;
+
+  public CouldNotResolveNameException(final String message) {
+    super(message);
+  }
+
+}