+++ pocs/poc-springboot-web/src/main/java/app5/persistence/primitives/HistoricPriceDAO.java
@@ -115,8 +115,8 @@ public class HistoricPriceDAO implements Serializable, ApplicationContextAware {
       return new SelectParentProductFromProductIdPhase(this.vo);
     }
 
-    public SelectParentProductFromSkuPhase fromSku() {
-      return new SelectParentProductFromSkuPhase(this.vo);
+    public SelectParentProductFromProp2uPhase fromProp2u() {
+      return new SelectParentProductFromProp2uPhase(this.vo);
     }
 
   }
@@ -135,16 +135,16 @@ public class HistoricPriceDAO implements Serializable, ApplicationContextAware {
 
   }
 
-  public class SelectParentProductFromSkuPhase {
+  public class SelectParentProductFromProp2uPhase {
 
     private HistoricPriceVO vo;
 
-    SelectParentProductFromSkuPhase(final HistoricPriceVO vo) {
+    SelectParentProductFromProp2uPhase(final HistoricPriceVO vo) {
       this.vo = vo;
     }
 
-    public ProductVO toSku() {
-      return productDAO.selectByUISku(this.vo.sku);
+    public ProductVO toProp2u() {
+      return productDAO.selectByUIProp2u(this.vo.prop2u);
     }
 
   }
@@ -198,8 +198,8 @@ public class HistoricPriceDAO implements Serializable, ApplicationContextAware {
     FROM_DATE$DESC("historic_price", "from_date", false), //
     PRICE("historic_price", "price", true), //
     PRICE$DESC("historic_price", "price", false), //
-    SKU("historic_price", "sku", true), //
-    SKU$DESC("historic_price", "sku", false);
+    PROP2U("historic_price", "sku", true), //
+    PROP2U$DESC("historic_price", "sku", false);
 
     private HistoricPriceOrderBy(final String tableName, final String columnName,
         boolean ascending) {
@@ -243,7 +243,7 @@ public class HistoricPriceDAO implements Serializable, ApplicationContextAware {
     public NumberColumn productId;
     public DateTimeColumn fromDate;
     public NumberColumn price;
-    public NumberColumn sku;
+    public NumberColumn prop2u;
 
     // Constructors
 
@@ -263,7 +263,7 @@ public class HistoricPriceDAO implements Serializable, ApplicationContextAware {
       this.productId = new NumberColumn(this, "product_id", "productId");
       this.fromDate = new DateTimeColumn(this, "from_date", "fromDate");
       this.price = new NumberColumn(this, "price", "price");
-      this.sku = new NumberColumn(this, "sku", "sku");
+      this.prop2u = new NumberColumn(this, "sku", "prop2u");
     }
 
   }